refactor: remove circular dependency
This introduces invisiblyDelete and deleteUserBucketEntry as modules
that can by used without having to require bucketDeletion as a whole.
This is required to avoid circular dependency when abortMultipartUpload
module is required within bucketDeletion.js
