+++ lib/api/apiUtils/bucket/bucketDeletion.js
@@ -3,51 +3,12 @@ const async = require('async');
 const { errors } = require('arsenal');
 
 const abortMultipartUpload = require('../object/abortMultipartUpload');
-const logger = require('../../../utilities/logger');
 
-const { usersBucket, oldUsersBucket, splitter, oldSplitter, mpuBucketPrefix } =
+const { splitter, oldSplitter, mpuBucketPrefix } =
     require('../../../../constants');
-const createKeyForUserBucket = require('./createKeyForUserBucket');
 const metadata = require('../../../metadata/wrapper');
 const kms = require('../../../kms/wrapper');
-
-function _deleteUserBucketEntry(bucketName, canonicalID, log, cb) {
-    log.trace('deleting bucket name from users bucket', { method:
-        '_deleteUserBucketEntry' });
-    const keyForUserBucket = createKeyForUserBucket(canonicalID, splitter,
-        bucketName);
-    metadata.deleteObjectMD(usersBucket, keyForUserBucket, {}, log, error => {
-        // If the object representing the bucket is not in the
-        // users bucket just continue
-        if (error && error.NoSuchKey) {
-            return cb(null);
-        // BACKWARDS COMPATIBILITY: Remove this once no longer
-        // have old user bucket format
-        } else if (error && error.NoSuchBucket) {
-            const keyForUserBucket2 = createKeyForUserBucket(canonicalID,
-                oldSplitter, bucketName);
-            return metadata.deleteObjectMD(oldUsersBucket, keyForUserBucket2,
-                {}, log, error => {
-                    if (error && !error.NoSuchKey) {
-                        log.error('from metadata while deleting user bucket',
-                            { error });
-                        return cb(error);
-                    }
-                    log.trace('deleted bucket from user bucket',
-                    { method: '_deleteUserBucketEntry' });
-                    return cb(null);
-                });
-        } else if (error) {
-            log.error('from metadata while deleting user bucket', { error,
-                method: '_deleteUserBucketEntry' });
-            return cb(error);
-        }
-        log.trace('deleted bucket from user bucket', {
-            method: '_deleteUserBucketEntry' });
-        return cb(null);
-    });
-}
-
+const deleteUserBucketEntry = require('./deleteUserBucketEntry');
 
 function _deleteMPUbucket(destinationBucketName, log, cb) {
     const mpuBucketName =
@@ -58,38 +19,6 @@ function _deleteMPUbucket(destinationBucketName, log, cb) {
             return cb();
         }
         return cb(err);
-    });
-}
-
-/**
- * Invisibly finishes deleting a bucket that already has a deleted flag
- * by deleting the object in the users bucket representing the created bucket
- * and then deleting the bucket in metadata
- * @param {string} bucketName - name of bucket
- * @param {string} canonicalID - bucket owner's canonicalID
- * @return {undefined}
- */
-function invisiblyDelete(bucketName, canonicalID) {
-    const log = logger.newRequestLogger();
-    log.trace('deleting bucket with deleted flag invisibly', { bucketName });
-    return _deleteUserBucketEntry(bucketName, canonicalID, log, err => {
-        if (err) {
-            log.error('error invisibly deleting bucket name from user bucket',
-            { error: err });
-            return log.end();
-        }
-        log.trace('deleted bucket name from user bucket');
-        return metadata.deleteBucket(bucketName, log, error => {
-            log.trace('deleting bucket from metadata',
-            { method: 'invisiblyDelete' });
-            if (error) {
-                log.error('error deleting bucket from metadata', { error });
-                return log.end();
-            }
-            log.trace('invisible deletion of bucket succeeded',
-            { method: 'invisiblyDelete' });
-            return log.end();
-        });
     });
 }
 
@@ -182,7 +111,7 @@ function deleteBucket(authInfo, bucketMD, bucketName, canonicalID, log, cb) {
         },
         function deleteUserBucketEntryStep(next) {
             log.trace('deleting bucket name from user bucket');
-            return _deleteUserBucketEntry(bucketName, canonicalID, log, next);
+            return deleteUserBucketEntry(bucketName, canonicalID, log, next);
         },
     ],
     // eslint-disable-next-line prefer-arrow-callback
@@ -206,7 +135,4 @@ function deleteBucket(authInfo, bucketMD, bucketName, canonicalID, log, cb) {
     });
 }
 
-module.exports = {
-    invisiblyDelete,
-    deleteBucket,
-};
+module.exports = deleteBucket;
