+++ lib/api/apiUtils/bucket/bucketShield.js
@@ -1,4 +1,4 @@
-const { invisiblyDelete } = require('./bucketDeletion');
+const invisiblyDelete = require('./invisiblyDelete');
 
 /**
  * Checks whether to proceed with a request based on the bucket flags
