+++ lib/api/bucketDelete.js
@@ -1,7 +1,7 @@
 const { errors } = require('arsenal');
 
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
-const { deleteBucket } = require('./apiUtils/bucket/bucketDeletion');
+const deleteBucket = require('./apiUtils/bucket/bucketDeletion');
 const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const { pushMetric } = require('../utapi/utilities');
